fix test for rename
